test(ng-primitives): address code review feedback
- guard menu item arrow-key navigation when the item is disabled
- keep progress aria-valuetext consistent with the clamped aria-valuenow
- exclude disabled items from becoming the roving-focus tab stop
- simplify checkbox/switch required to a plain input (not controlled)
- extract the shared SpeechRecognition mock out of a test file
- add a missing detectChanges and tighten a few test names
